Rename login route file to index
